+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 * New: Make hiding main window during screenshot optional. ([#386](https://github.com/ksnip/ksnip/issues/386))
 * New: Open several files at once in tabs. ([#355](https://github.com/ksnip/ksnip/issues/355))
 * New: Allow modifying size of selected rectangle before making screenshot. ([#197](https://github.com/ksnip/ksnip/issues/197))
+* New: Option to keep main window hidden after a taking screenshot. ([#409](https://github.com/ksnip/ksnip/issues/409))
 * New kImageAnnotator: Add Pixelate image area tool. ([#140](https://github.com/ksnip/kImageAnnotator/issues/140))
 * New kImageAnnotator: Zoom in and out. ([#123](https://github.com/ksnip/kImageAnnotator/issues/123))
 * New kImageAnnotator: Add interface for adding custom tab context menu actions. ([#96](https://github.com/ksnip/kImageAnnotator/issues/96))